Add excludes feature to allow coverage excludes
The coverage task did not support GRCOV_EXCL_xxx markers that
will help with marking those areas not intended to have coverage.

This feature adds support for the tags:

GRCOV_EXCL_LINE
GRCOV_EXCL_START
GRCOV_EXCL_STOP

It was added as a feature so that it didn't accidentally change
the behavior of someone using this coverage task already.  You
need to opt in by specifying "excludes" to have those tags added
on the command line when the xtask coverage command executes.